Enable linux build and CI/CD (#2)
This adds minor changes to the codebase to allow it to compile and run
on a Linux environment. Additionally, this adds a CMakeLists.txt file
which can compile the project on Linux, and also adds a Github Actions
script which performs the build/test run on each push to the repository.