Updating evaluation library, adding tests (#3)
This incorporates changes from the
[pokermania/pokereval](https://github.com/pokermania/pokereval)
repository, which as far as I'm able to tell appears to be the best
location to obtain the base poker-eval repository from. This also adds
poker-eval tests to the CI/CD environment on each push, allowing not
only for the overall Poker Equity Test to be ran, but also for the
poker-eval tests to be ran. Along with that I have a few changes I made:
* I couldn't directly just copy/paste the other poker-eval repository as
it looks like some t_ files need to be generated and I'm uncertain how
to do that. So instead, I went file-by-file and incorporated the changes
into the files that appeared to need it
* I'm not using the automake files it came with but instead using the
(much more user friendly) CMakeLists.txt files that I've added to the
repository. I've replicated rebuilding the project using these cmake
files
* Tests from `Poker.PokerEval/tests/run.in` and
`Poker.PokerEval/examples/utest1` have been incorporated into the CI/CD
flow
* Some tests in `Poker.PokerEval/examples/utest1` did not run, either
due to incorrect command-line arguments or the test program reporting
that the requested test is not supported. These tests have been omitted
* There are tests in `Poker.PokerEval/tests/bug1823` that are not
incorporated here
* I've modified the file `Poker.PokerEval/tests/razz.c` as it appears as
though the testing logic for the second case was in reverse and thus
always failed. I am personally unfamiliar with the rules of Razz poker,
so I used [an external
website](https://www.cardplayer.com/poker-tools/odds-calculator/razz) to
enter the hands and verify that the output is correct. It looks like the
library was correctly evaluating the result, but the test was
incorrectly checking the return.
* License has been updated to GPLv3, which is the updated license from
the newer codebase
* The tools `evx_gen5` and `evx_gen7` do not appear to successfully
compile. I've commented out those builds but left them in the overall
CMakeLists.txt file to look into in the future
* I had to modify the `Poker.PokerEval/tests/poker_wrapper.c` file to
add the wrapper functions which were not present in the above referenced
repository. However, I found [another
repository](https://github.com/dooglus/pokereval/blob/920dc2d7081f7d58647853d94b860015258916ac/lib/poker_wrapper.c)
which had definitions for those functions